Loading spinner for submission history table (#3640)
* Loading spinner for submission history table

* Lint

* Revert proxy change

* Handle null sub history row

* Handle null pks for sub history
